move HttpError to http module
